+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project will be documented in this file.
 
+## [3.1.2] - 2024-02-05
+- Added script *DisableOtherCameras.cs* to deactivate all Cameras except the Rig Camera on scene load.
+
 ## [3.1.0] - 2023-10-16
 - Added dynamic handmenu panels which can be placed in a scene and get loaded when the scene is entered and unloaded when leaving the scene
 - Minor restructure of mainmenu layout